feature/refactoring, разнес скрипты по папкам
