+++ folder-tree.pgsql
@@ -1,0 +1,25 @@
+giga-chat-extension
+├── create-gigachat-plugin.sh
+├── folder-tree.pgsql
+├── media
+├── node_modules
+├── out
+│   ├── extension.js
+│   ├── fetchGigaChainSuggestion.js
+│   ├── fetchGigaChatSuggestion.js
+│   ├── fetchLocalSuggestion.js
+│   ├── gigaChainInlineCompletion.js
+│   ├── inlineCompletion.js
+│   ├── output.js
+│   └── stubSuggestion.js
+├── package-lock.json
+├── package.json
+├── src
+│   ├── extension.ts
+│   ├── fetchGigaChainSuggestion.ts
+│   ├── fetchGigaChatSuggestion.ts
+│   ├── gigaChainInlineCompletion.ts
+│   ├── output.ts
+│   └── stubSuggestion.ts
+├── synthetics
+└── tsconfig.json
