+++ src/core/output.ts
