+++ src/core/stubSuggestion.ts
